+++ dummy/src/main.rs
@@ -1,10 +1,10 @@
 use nutype::nutype;
 
 #[nutype(
-    validate(greater = 0, less = 24, predicate = |x| x == &4),
-    derive(Debug, PartialEq, Deref, AsRef, Default),
-    default = 4
+    validate(predicate = |v| v),
+    derive(Default),
+    default = false
 )]
-pub struct Hour(i32);
+pub struct TestData(bool);
 
 fn main() {}