+++ nutype_macros/src/common/gen/tests.rs
@@ -1,5 +1,5 @@
 use proc_macro2::TokenStream;
-use quote::quote;
+use quote::{quote, ToTokens};
 
 use crate::common::models::{NumericBound, TypeName};
 
@@ -16,12 +16,49 @@ where
         return None;
     };
 
-    let msg = format!("\nInconsistent lower and upper boundaries for type `{type_name}`\nThe upper boundary `{upper}` must be greater than or equal to the lower boundary `{lower}`\n");
+    let msg = format!(
+        "
+Inconsistent lower and upper boundaries for type `{type_name}`
+The upper boundary `{upper}` must be greater than or equal to the lower boundary `{lower}`
+Note: the test is generated automatically by #[nutype] macro.
+"
+    );
 
     Some(quote!(
         #[test]
         fn should_have_consistent_lower_and_upper_boundaries() {
             assert!(#upper >= #lower, #msg);
+        }
+    ))
+}
+
+pub fn gen_test_should_have_valid_default_value(
+    type_name: &TypeName,
+    maybe_default_value: &Option<syn::Expr>,
+    has_validation: bool,
+) -> Option<TokenStream> {
+    if !has_validation {
+        // If there is no validation, then every possible default value will be valid,
+        // so there is no need to generate the test.
+        return None;
+    }
+
+    let default_value: TokenStream = maybe_default_value.as_ref()?.to_token_stream();
+
+    let msg = format!(
+        "
+Type `{type_name}` has invalid default value `{default_value}`
+Note: the test is generated automatically by #[nutype] macro
+"
+    );
+
+    Some(quote!(
+        #[test]
+        fn should_have_valid_default_value() {
+            let default_inner_value = #type_name::default().into_inner();
+            // Typically `::default()` would already panic, but in case if the panic is removed by
+            // some reason we still want the test to fail.
+            #type_name::new(default_inner_value).expect(#msg);
         }
     ))
 }