+++ nutype_macros/src/integer/gen/mod.rs
@@ -17,7 +17,11 @@ use super::{
 use crate::common::{
     gen::{
         error::gen_error_type_name,
-        tests::gen_test_should_have_consistent_lower_and_upper_boundaries, traits::GeneratedTraits,
+        tests::{
+            gen_test_should_have_consistent_lower_and_upper_boundaries,
+            gen_test_should_have_valid_default_value,
+        },
+        traits::GeneratedTraits,
         GenerateNewtype,
     },
     models::{ErrorTypeName, Guard, TypeName},
@@ -143,15 +147,23 @@ where
     fn gen_tests(
         type_name: &TypeName,
         _inner_type: &Self::InnerType,
-        _maybe_default_value: &Option<syn::Expr>,
+        maybe_default_value: &Option<syn::Expr>,
         guard: &Guard<Self::Sanitizer, Self::Validator>,
+        _traits: &HashSet<Self::TypedTrait>,
     ) -> TokenStream {
         let test_lower_vs_upper = guard.validators().and_then(|validators| {
             gen_test_should_have_consistent_lower_and_upper_boundaries(type_name, validators)
         });
 
+        let test_valid_default_value = gen_test_should_have_valid_default_value(
+            type_name,
+            maybe_default_value,
+            guard.has_validation(),
+        );
+
         quote! {
             #test_lower_vs_upper
+            #test_valid_default_value
         }
     }
 }